mock out makerequest
