+++ .gitignore
@@ -16,3 +16,4 @@ Thumbs.db
 drone-openapi
 coverprofile.out
 testdata/specs/hello_world_spec.json
+testdata/specs/users_spec.json
